check for internet now separate from check for multiplayerServer removed frunctions from PlatformSpecificHelpers__
