+++ core/src/com/unciv/ui/utils/GeneralPlatformSpecificHelpers.kt
@@ -13,8 +13,6 @@ interface GeneralPlatformSpecificHelpers {
      */
     fun allowPortrait(allow: Boolean) {}
 
-    fun isInternetConnected(): Boolean
-
     /**
      * Notifies the user that it's their turn while the game is running
      */
